fix: configure the project to work with taskcat
Configure the project to work with the AWS CDK, AWS SAM, and taskcat.
Because the @aws-cdk/aws-lambda module generates random CloudFormation
parameter names for the S3 bucket and key names when uploading functions
from an asset directory, they cannot be predicted during a CI/CD build,
and taskcat cannot be leveraged to run deployment tests, as it expects
the parameter names to be static (see https://git.io/JMB4L for more
details).

To work around this, we create a runtime context value called
isTaskcatRun (see https://go.aws/3nZWRCp for more details). If that
value is set to true, we define static parameter names for the S3 bucket
and key names in the CloudFormation template, and leverage taskcat's
ability to bundle Lambda functions to zip and upload the code to S3.

To conform to the folder structure expected by taskcat for packaging
Lambda functions, we move the function's code to a subdirectory under
"lambda/". Each directory under that folder is considered as a single
Lambda function by taskcat, and is zipped separately.