feat: implement SES password generation logic
Implement the SES password generation logic using the pseudocode and
example Python implementation made available at https://go.aws/3rE70XN.

To help write the unit tests, we pass the example secret access key
provided in the docs to the Python script, and validate that both our
Lambda function and the Python script return the same value.

This also adds the "NoEcho" parameter in the response, to make sure that
the SES password is masked when making API calls or retrieving the value
through functions like Fn::GetAtt, for example.